requirements.txt
facebook-sdk==2.0.0
py==1.4.31
pytest==3.0.2
requests==2.11.1
